+++ api_methods.py
@@ -1,11 +1,33 @@
 import requests
 import json
-import logging
 import allure
 import endpoints
-from faker import Faker
 import random
-import string
+from utilites import Funcs, Generator
+
+
+class Request:
+
+    @staticmethod
+    @allure.step('POST request.')
+    def post(path, payload, access_token):
+        return requests.post(endpoints.URL_SERVICE + path, headers=Funcs.auth_header(access_token), data=json.dumps(payload))
+
+    @staticmethod
+    @allure.step('DELETE request.')
+    def delete(path, access_token):
+        return requests.delete(endpoints.URL_SERVICE + path, headers=Funcs.auth_header(access_token))
+
+    @staticmethod
+    @allure.step('PATCH request.')
+    def patch(path, access_token, payload):
+        return requests.patch(endpoints.URL_SERVICE + path, headers=Funcs.auth_header(access_token), data=json.dumps(payload))
+
+    @staticmethod
+    @allure.step('GET request.')
+    def get(path, access_token):
+        return requests.get(endpoints.URL_SERVICE + path, headers=Funcs.auth_header(access_token))
+
 
 class User:
 
@@ -30,6 +52,7 @@ class User:
             Funcs.message(f'Пользователь удален')
         else:
             Funcs.message(f'Пользователь не удален: {response.status_code}, {response.text}.')
+
 
 class Order:
 
@@ -96,83 +119,3 @@ class Order:
             return Order.nums_list(r)
         else:
             Funcs.message(f'Список заказов не получен: {response.status_code}, {response.text}.')
-
-
-class Request:
-
-    @staticmethod
-    @allure.step('POST request.')
-    def post(path, payload, access_token):
-        return requests.post(endpoints.URL_SERVICE + path, headers=Funcs.auth_header(access_token), data=json.dumps(payload))
-
-    @staticmethod
-    @allure.step('DELETE request.')
-    def delete(path, access_token):
-        return requests.delete(endpoints.URL_SERVICE + path, headers=Funcs.auth_header(access_token))
-
-    @staticmethod
-    @allure.step('PATCH request.')
-    def patch(path, access_token, payload):
-        return requests.patch(endpoints.URL_SERVICE + path, headers=Funcs.auth_header(access_token), data=json.dumps(payload))
-
-    @staticmethod
-    @allure.step('GET request.')
-    def get(path, access_token):
-        return requests.get(endpoints.URL_SERVICE + path, headers=Funcs.auth_header(access_token))
-
-
-class Funcs:
-
-    @staticmethod
-    def message(message):
-        allure.attach('', message, attachment_type=allure.attachment_type.TEXT)
-        logging.info(message)
-
-    @staticmethod
-    @allure.step('Изменение одного из полей данных для соответствия тестовому сценарию.')
-    def change_payload_param(param, key, value):
-        param[key] = value
-        return param
-
-    @staticmethod
-    @allure.step('Oбновление токенов в данных пользователя')
-    def update_user_tokens(user, response):
-        user["accessToken"] = response.json()["accessToken"]
-        user["refreshToken"] = response.json()["refreshToken"]
-
-    @staticmethod
-    @allure.step('Создание payload  для логина из данных зарегистрированного пользователя.')
-    def login_payload(user):
-        payload = {
-            "email":f'{user["user"]["email"]}',
-            "password": f'{user["user"]["password"]}'
-            }
-        return payload
-
-    @staticmethod
-    @allure.step('Создание header c токеном авторизации.')
-    def auth_header(token):
-        header = {
-            "Content-Type": "application/json",
-            'Authorization': f'{token}'
-            }
-        return header
-
-
-class Generator:
-
-    @staticmethod
-    @allure.step('Генерация уникальных данных пользователя.')
-    def get_user_payload():
-        fake = Faker()
-        return {
-            "email": f'"{fake.email()}"',
-            "password": f'"{fake.password()}"',
-            "name": ''.join(random.choice(string.ascii_lowercase) for i in range(8)).capitalize()
-            }
-
-    @staticmethod
-    @allure.step('Генерация хеша ингредиентов.')
-    def hash():
-        fake = Faker()
-        return fake.md5()
