+++ tests/test_user_registration.py
@@ -1,8 +1,9 @@
 import allure
 import endpoints
 from data import Message
-from utils import Funcs, Request, Generator
 import pytest
+from utilites import Funcs, Generator
+from api_methods import Request
 
 class TestUserRegistration:
